+++ srcs/headers/primitives.h
@@ -19,9 +19,9 @@
 
 typedef struct s_float3
 {
-	float		x;
-	float		y;
-	float		z;
+	double		x;
+	double		y;
+	double		z;
 }	t_float3;
 
 typedef struct s_int3
@@ -51,28 +51,28 @@ typedef struct s_ray
 typedef struct s_material
 {
     t_color    	color; //hex
-    float       specular_strength; // Specular highlight strength
-    float       reflectivity; // 0 (not reflective) to 1 (perfect mirror)
-	float       roughness; // 0 (perfectly smooth) to 1 (rough)
+    double       specular_strength; // Specular highlight strength
+    double       reflectivity; // 0 (not reflective) to 1 (perfect mirror)
+	double       roughness; // 0 (perfectly smooth) to 1 (rough)
 }   t_material;
 
 typedef struct s_hit
 {
     t_point     point;
     t_vector    normal;
-    float       distance;
+    double       distance;
     t_material	*material;
 }   t_hit;
 
 t_vector	vec_add(const t_vector a, const t_vector b);
 t_vector	vec_add(const t_vector a, const t_vector b);
 t_vector	vec_sub(const t_vector a, const t_vector b);
-t_vector	vec_scale(const float scalar, const t_vector v);
+t_vector	vec_scale(const double scalar, const t_vector v);
 t_vector	vec_cross(const t_vector a, const t_vector b);
 t_vector	vec_normalize(const t_vector v);
 t_vector	vec_negate(const t_vector a);
-float		vec_dot(const t_vector a, const t_vector b);
-float		vec_length(const t_vector a);
+double		vec_dot(const t_vector a, const t_vector b);
+double		vec_length(const t_vector a);
 bool 		is_vec_equal(const t_vector a, const t_vector b);
